third_party/blink/web_tests/external/wpt/subresource-integrity/integrity-policy/script.https-expected.txt

This is a testharness.js-based test.
[FAIL] Ensure that an about:blank URL script with no integrity runs
  assert_equals: No policy violation report should be observed expected null but got object "[object Object]"
[FAIL] Ensure that a no-CORS about:blank URL script with no integrity runs
  assert_equals: No policy violation report should be observed expected null but got object "[object Object]"
Harness: the test ran to completion.
